Rename: select와 search와 card로 scss 파일 분리
