+++ src/components/cardMaker/cards.module.scss
@@ -50,111 +50,6 @@
   display: flex;
 }
 
-.pokemon_select_wrapper {
-  display: flex;
-  flex-direction: column;
-  border-radius: 10px;
-  border: 1px solid #ebebeb;
-  background-color: #fff;
-  padding: 17px 21px;
-  justify-content: space-between;
-  width: fit-content;
-  gap: 10px;
-
-  div {
-    display: flex;
-    gap: 15px;
-  }
-}
-
-.pokemon_select_wrapper > div > .border_button {
-  width: 120px;
-  height: 204px;
-  border-radius: 15px;
-  border: 3px solid #0f60ce;
-  font-size: 21px;
-  font-weight: 500;
-}
-
-.pokemon_select_wrapper > div > .page_button {
-  cursor: pointer;
-  width: 32px;
-  border: none;
-  outline: none;
-  border-radius: 10px;
-  color: #b5b5b5;
-  background-color: #f8f8f8;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 20px;
-}
-
-.pokemon_select_wrapper > div > .title {
-  font-size: 28px;
-  font-weight: 500;
-}
-
-.border_button {
-  cursor: pointer;
-  outline: none;
-  color: #0f60ce;
-  border: 2px solid #0f60ce;
-  background-color: #fff;
-  color: #0f60ce;
-  font-weight: 500;
-  font-family: GmarketSansMedium;
-  transition: 0.2s ease;
-
-  &:hover {
-    background-color: #0f60ce;
-    color: #fff;
-  }
-}
-
-.cards {
-  &__row_cards {
-    display: flex;
-    gap: 15px;
-
-    &__column {
-      gap: 15px;
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-    }
-
-    &__card {
-      width: 121px;
-      height: 121px;
-      border-radius: 15px;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      border: 2.3px solid #e5e5e5;
-      background: #fff;
-
-      img {
-        width: 100px;
-        height: 100px;
-      }
-    }
-
-    span {
-      font-size: 18px;
-      font-weight: 400;
-      margin-bottom: 10px;
-    }
-
-    button {
-      width: 99px;
-      height: 34px;
-      border-radius: 8px;
-      font-size: 13px;
-    }
-  }
-}
-
 .editor_wrapper {
   justify-content: space-between;
   display: flex;
@@ -249,127 +144,6 @@
 
 .list_name {
   font-size: 18px;
-}
-
-.search_form_wrapper {
-  width: 100%;
-  height: 100px;
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  position: relative;
-  align-items: flex-start;
-}
-
-.search__wrapper {
-  z-index: 3;
-  position: absolute;
-  top: 40px;
-  left: 0;
-  background-color: #fff;
-  padding: 15px 20px;
-  min-height: 71px;
-  border-radius: 15px;
-  border: 2.3px solid #000;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: 20px;
-}
-
-.search__inner {
-  display: flex;
-  width: 365px;
-  height: 36px;
-  border-bottom: 1px solid #000;
-}
-
-.search__input {
-  width: 100%;
-  outline: none;
-  border: none;
-  background-color: transparent;
-}
-
-.search__button {
-  outline: none;
-  background-color: transparent;
-  border: none;
-}
-
-.dropdown_wrapper {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 20px;
-}
-
-.dropdown__container {
-  display: grid;
-  grid-template-columns: repeat(3, minmax(110px, 1fr));
-  gap: 10px;
-}
-
-.dropdown__list {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 7px;
-
-  div {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border-radius: 4px;
-    border: 1px solid #e3e3e3;
-    width: 110px;
-    height: 110px;
-
-    img {
-      width: 80%;
-    }
-  }
-
-  span {
-    font-size: 18px;
-    font-weight: 400;
-  }
-}
-
-.pagination_wrapper {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 10px;
-}
-
-.pagination_button {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-  border: none;
-  outline: none;
-  background-color: transparent;
-  width: 23px;
-  height: 20px;
-}
-
-.button__action {
-  font-size: 18px;
-  font-weight: 700;
-}
-
-.arrow_button {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background-color: transparent;
-  width: 30px;
-  height: 30px;
-  border: 1px solid;
-  border-radius: 10px;
 }
 
 .save_button {
