+++ src/components/cardMaker/searchPokemon/search.module.scss
@@ -1,0 +1,120 @@
+.search_form_wrapper {
+  width: 100%;
+  height: 100px;
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  position: relative;
+  align-items: flex-start;
+}
+
+.search__wrapper {
+  z-index: 3;
+  position: absolute;
+  top: 40px;
+  left: 0;
+  background-color: #fff;
+  padding: 15px 20px;
+  min-height: 71px;
+  border-radius: 15px;
+  border: 2.3px solid #000;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: 20px;
+}
+
+.search__inner {
+  display: flex;
+  width: 365px;
+  height: 36px;
+  border-bottom: 1px solid #000;
+}
+
+.search__input {
+  width: 100%;
+  outline: none;
+  border: none;
+  background-color: transparent;
+}
+
+.search__button {
+  outline: none;
+  background-color: transparent;
+  border: none;
+}
+
+.dropdown_wrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 20px;
+}
+
+.dropdown__container {
+  display: grid;
+  grid-template-columns: repeat(3, minmax(110px, 1fr));
+  gap: 10px;
+}
+
+.dropdown__list {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 7px;
+
+  div {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    border-radius: 4px;
+    border: 1px solid #e3e3e3;
+    width: 110px;
+    height: 110px;
+
+    img {
+      width: 80%;
+    }
+  }
+
+  span {
+    font-size: 18px;
+    font-weight: 400;
+  }
+}
+
+.pagination_wrapper {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 10px;
+}
+
+.pagination_button {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  cursor: pointer;
+  border: none;
+  outline: none;
+  background-color: transparent;
+  width: 23px;
+  height: 20px;
+}
+
+.button__action {
+  font-size: 18px;
+  font-weight: 700;
+}
+
+.arrow_button {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background-color: transparent;
+  width: 30px;
+  height: 30px;
+  border: 1px solid;
+  border-radius: 10px;
+}
