+++ src/components/cardMaker/selectPokemon/select.module.scss
@@ -1,0 +1,104 @@
+.cards {
+  &__row_cards {
+    display: flex;
+    gap: 15px;
+
+    &__column {
+      gap: 15px;
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+    }
+
+    &__card {
+      width: 121px;
+      height: 121px;
+      border-radius: 15px;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      border: 2.3px solid #e5e5e5;
+      background: #fff;
+
+      img {
+        width: 100px;
+        height: 100px;
+      }
+    }
+
+    span {
+      font-size: 18px;
+      font-weight: 400;
+      margin-bottom: 10px;
+    }
+
+    button {
+      width: 99px;
+      height: 34px;
+      border-radius: 8px;
+      font-size: 13px;
+    }
+  }
+}
+
+.pokemon_select_wrapper {
+  display: flex;
+  flex-direction: column;
+  border-radius: 10px;
+  border: 1px solid #ebebeb;
+  background-color: #fff;
+  padding: 17px 21px;
+  justify-content: space-between;
+  width: fit-content;
+  gap: 10px;
+
+  div {
+    display: flex;
+    gap: 15px;
+  }
+}
+
+.pokemon_select_wrapper > div > .border_button {
+  width: 120px;
+  height: 204px;
+  border-radius: 15px;
+  border: 3px solid #0f60ce;
+  font-size: 21px;
+  font-weight: 500;
+}
+
+.pokemon_select_wrapper > div > .page_button {
+  cursor: pointer;
+  width: 32px;
+  border: none;
+  outline: none;
+  border-radius: 10px;
+  color: #b5b5b5;
+  background-color: #f8f8f8;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 20px;
+}
+
+.pokemon_select_wrapper > div > .title {
+  font-size: 28px;
+  font-weight: 500;
+}
+
+.border_button {
+  cursor: pointer;
+  outline: none;
+  color: #0f60ce;
+  border: 2px solid #0f60ce;
+  background-color: #fff;
+  color: #0f60ce;
+  font-weight: 500;
+  font-family: GmarketSansMedium;
+  transition: 0.2s ease;
+
+  &:hover {
+    background-color: #0f60ce;
+    color: #fff;
+  }
+}
